added refs to block and chain data to control re-renders and stale blocks
